refactor: answers routing
